kb: Support 'normalizing' header depths when including subdocuments.
The top-level header of a document is always h1, even if all the headers in
the document itself are h2 or h3.  But you can now [[include+2:refname]] to
increase all the headers in the included document by 2 levels, so it can be
included as a whole chapter inside an h2.