kb: base depth should be h3 for now.
This is kind of lame.  The top-level title in our template is currently h2,
so any subtitles in a document need to be lower than that.  Really, it's our
template that's wrong, at least off by one. (So the top level title should
be h1 instead.)

Anyway, it's easy enough to change later.