Read title, tags, and mtime from the kb doc file every time.
We were already loading the text from the file instead of the database, but
that was weird because updating tags, title, and mtime didn't work.  Now
they're all always up to date, which is better when you're editing things
live.  It doesn't actually *save* the results though, by default.  It also
still doesn't update the search index, but I think that's reasonable.