Don't update Doc.title unless we call Doc.use_latest().
The document viewer should get the latest of everything, but it's confusing
when you search on a particular keyword and the title is different. This is
especially true when searching for "duplicate" to find articles that had to
be renamed to whatever[duplicate] because 'whatever' was already taken: the
search engine would find the doc, but not display the [duplicate] marker in
the title, since it wasn't in the file, only in the database.