Even shorter nicedate() for article postinfo.
Now we give the actual date if it was more than 14 days ago, and just say
"minutes ago" or "seconds ago" or "today" if those thing are true.  This
saves a little bit of horizontal space, which helps reduce word wrap for
articles that are in several categories.

Plus less clutter is better, all else being equal.