latex renderer now renders books and articles.
A "book" has headers of the form:

Where "###" is the depth of the section, "title text" is the actual section
name, and "filename" is the file that contains the (first part of) the
content for that section.

In latex mode, we convert the above into essentially:

	### title text

	[[include+3:filename]]

(where '3' is the header depth).  The result is that we can generate an
entire book from a single file that just contains hyperlink-headings.

In normal html browser mode, we don't change the original form; it's valid
markdown that simply makes a header that's a hyperlink.  A bunch of these in
a row is just a bunch of hyperlinks to book sections.