ekb/models: add a new kind of [[include]] expansion.
Now we support [[include:doc-file-name#Section Name]].  This basically
deletes everything in the included text up to and including the "## Section
Name" header, then pastes in only the stuff that came after.

This is pretty handy for including FAQ or KB type articles in a book.  The
article title, "Question" subtitle, question itself, and "Answer" subtitle
are kind of redundant with your book's subsection title.  But you still want
to be able to include everything after the "Answer" subtitle.  To do it, you
just [[include:my-kb-article#Answer]].

This also works in "book style" inclusion syntax, which will look something
like:  ### [This is a title][my-kb-article#Answer]