Instead of optipng, use the python-imaging (PIL) library.
optipng is cool and makes smaller files, but it's really very slow compared
to PIL.  It's also one more dependency we need in order to run.

It might make more sense to *optionally* allow optipng, and to spin it off
in the background to optimize images after they're uploaded or something.
But for now, too slow.

This makes rendering an image-laden document into pdf (for the first time)
fast again.