Initial implementation of an image uploading feature in edit.html.
You can now upload image files, then click a link next to each image to
insert a link to it into the text.

The uploading is AJAX-enabled and rather nifty.

The page layout, however, leaves a bit to be desired (especially on small
screens).  It would also be nice if it auto-grabbed the set of images
already link from this file, and added them to the list by default.

Also, we should remove the need for all image filenames to have a
/static/kbimages/ prefix in their name; that's an implementation detail, not
something that should actually end up in the articles themselves.  Oh well.