Clean up the css for img and blockquote.
Blockquotes are usually used for "Note: whatever" in our implementation, so
it should be less italic and more formal.

img objects are typically screenshots, so there's no need to decorate them
with an extra border; screenshots are often decorated as it is, so that they
look good when printed.  (Arguably, maybe we should auto-decorate the images
so that plain screenshots will look better.  But for now we won't.)