Track which articles refer to which other articles.
This makes it a little easier to find your way around when you're viewing an
article that's used in a particular book: it'll give you a link to that book
(or at least to the chapter that includes this article, and if you go to
that chapter, it'll have a link to the book).

Not perfect, UI-wise, since it really ought to be displayed as a hierarchy
or something.  But this should hopefully help readers and editors find their
way around.  It'll also reveal when an article isn't used in *any* books.