edit.html: use soft linewraps instead of hard linewraps.
I would have liked it better if the saved text was nicely wrapped at 78
columns... but this turns out to be no good.  For example, if you have an
image filename that's more than 80 characters, it's just not going to work
if it gets wrapped onto more than one line, since markdown can't handle it.
Similar with long headings.